+++ packages/dumb/src/components/tooltips/warm-up-tooltip/warm-up-tooltip.tsx
@@ -24,10 +24,9 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-import { Tooltip } from '../../../index'
 import React, { FC } from 'react'
 import { TimePrefs, WarmUp } from 'medical-domain'
-import { DEFAULT_TOOLTIP_OFFSET, Position } from '../common/tooltip/tooltip'
+import { DEFAULT_TOOLTIP_OFFSET, Position, Tooltip } from '../common/tooltip/tooltip'
 import { getDateTitleForBaseDatum } from '../../../utils/tooltip/tooltip.util'
 import { TooltipLine } from '../common/tooltip-line/tooltip-line'
 import colors from '../../../styles/colors.css'