+++ packages/tideline/plugins/blip/chartdailyfactory.js
@@ -26,7 +26,7 @@ import { MGDL_UNITS } from 'medical-domain'
 import Pool from '../../js/pool'
 import oneDay from '../../js/oneday'
 import fill from '../../js/plot/util/fill'
-import { createYAxisBasal, createYAxisBG, createYAxisBolus } from '../../js/plot/util/scales'
+import { createYAxisBasal, createYAxisBG, createYAxisBolus, createYAxisIob } from '../../js/plot/util/scales'
 import axesDailyx from '../../js/plot/util/axes/dailyx'
 import plotZenModeEvent from '../../js/plot/zenModeEvent'
 import plotPhysicalActivity from '../../js/plot/physicalActivity'
@@ -45,7 +45,8 @@ import plotMessage from '../../js/plot/message'
 import plotTimeChange from '../../js/plot/timechange'
 import plotNightMode from '../../js/plot/nightModeEvent'
 import plotEventSuperposition from '../../js/plot/eventSuperposition'
-import { getDataWithoutSuperpositionEvents, getSuperpositionEvents } from 'dumb'
+import { getDataWithoutSuperpositionEvents, getSuperpositionEvents, isDBLG2 } from 'dumb'
+import plotIob from '../../js/plot/iob'
 
 /**
  * @typedef {import('../../js/tidelinedata').default } MedicalDataService
@@ -88,6 +89,11 @@ function chartDailyFactory(parentElement, tidelineData, options = {}) {
 
   chart.id(parentElement.id).width(width).height(height)
   d3.select(parentElement).call(chart)
+
+  const pumpSettings = tidelineData?.medicalData?.pumpSettings
+  const hasPumpSettings = pumpSettings?.length > 0
+  const isDblg2User = hasPumpSettings ? isDBLG2(pumpSettings[0].payload.device.name) : false
+
   // ***
   // Setup Pools
   // ***
@@ -148,7 +154,7 @@ function chartDailyFactory(parentElement, tidelineData, options = {}) {
       }],
       baseline: options.labelBaseline
     }])
-    .heightRatio(0.5)
+    .heightRatio(isDblg2User ? 0.5 : 0.4)
     .gutterWeight(1.0)
 
   // carbs and boluses data pool
@@ -209,6 +215,29 @@ function chartDailyFactory(parentElement, tidelineData, options = {}) {
     .heightRatio(1.0)
     .gutterWeight(1.0)
 
+
+  let poolIob = null
+  if (isDblg2User) {
+    poolIob = new Pool(chart)
+    chart.addPool(poolIob)
+    const poolIobId = 'poolIob'
+    poolIob
+      .id(poolIobId, chart.poolGroup)
+      .dataTestId('iob-section', poolIobId)
+      .labels([{
+        spans: [{
+          text: t('active-insulin'),
+          className: 'label-main'
+        }, {
+          text: ` (${t('U')})`,
+          className: 'label-light'
+        }],
+        baseline: options.labelBaseline
+      }])
+      .heightRatio(1.5)
+      .gutterWeight(1.5)
+  }
+
   chart.arrangePools()
   chart.setTooltip()
 
@@ -220,24 +249,6 @@ function chartDailyFactory(parentElement, tidelineData, options = {}) {
   chart.tooltips.addGroup(poolMessages, {
     type: 'message',
     shape: 'generic'
-  })
-  chart.tooltips.addGroup(poolBG, {
-    type: 'cbg',
-    classes: ['d3-bg-low', 'd3-bg-target', 'd3-bg-high']
-  })
-  chart.tooltips.addGroup(poolBG, {
-    type: 'smbg'
-  })
-  chart.tooltips.addGroup(poolBolus, {
-    type: 'wizard',
-    shape: 'generic'
-  })
-  chart.tooltips.addGroup(poolBolus, {
-    type: 'bolus',
-    shape: 'generic'
-  })
-  chart.tooltips.addGroup(poolBasal, {
-    type: 'basal'
   })
 
   // ***
@@ -294,7 +305,6 @@ function chartDailyFactory(parentElement, tidelineData, options = {}) {
     onPhysicalOut: options.onTooltipOut,
     tidelineData
   }))
-
 
   const eventSuperpositionItems = getSuperpositionEvents(tidelineData)
   const alarmEvents = getDataWithoutSuperpositionEvents(tidelineData.medicalData.alarmEvents, eventSuperpositionItems)
@@ -419,6 +429,31 @@ function chartDailyFactory(parentElement, tidelineData, options = {}) {
     onConfidentialHover: options.onConfidentialHover,
     onConfidentialOut: options.onTooltipOut
   }))
+
+  if (isDblg2User) {
+    // IOB pool
+    // setup axis & main y scale
+    poolIob?.axisScaleFn(createYAxisIob)
+
+    // add background fill rectangles to IOB pool
+    poolIob?.addPlotType({ type: 'fill' }, fill(poolIob, {
+      isDaily: true
+    }))
+
+    // add IOB curve to IOB pool
+    poolIob?.addPlotType({ type: 'iob' }, plotIob(poolIob, {
+      tidelineData,
+      onIobHover: options.onIobHover,
+      onIobOut: options.onTooltipOut
+    }))
+
+    // Add confidential mode to IOB pool: Must be the last in the pool to mask stuff below
+    poolIob?.addPlotType({ type: 'deviceEvent', name: 'confidential' }, plotConfidentialModeEvent(poolIob, {
+      tidelineData,
+      onConfidentialHover: options.onConfidentialHover,
+      onConfidentialOut: options.onTooltipOut
+    }))
+  }
 
   // messages pool
   // add background fill rectangles to messages pool