+++ packages/yourloops/components/patient-data/patient-data.utils.ts
@@ -193,6 +193,7 @@ export class PatientDataUtils {
       cbg: [],
       confidentialModes: [],
       deviceParametersChanges: [],
+      iob: [],
       messages: [],
       meals: [],
       nightModes: [],