+++ CHANGELOG.md
@@ -6,6 +6,7 @@ It is based on Tidepool Blip 1.27.
 ### Added
 - YLP-2344 Patient profile information page for HCPs only
 - YLP-3333 Tooltip for Zen mode in Daily view
+- YLP-3580 IOB chart for DBLG2 users in Daily view
 - YLP-3589 add new estimated total insulin statistic
 - YLP-3611 Manage events superposition in Daily view
 - YLP-3612 Make events superposition popover draggable